requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

wget==3.2
numpy==1.26.4
pillow==10.2.0
fastai==2.7.14
ipython==8.22.1
torch==2.2.1+cpu
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
torchvision==0.17.1+cpu
opencv-python==4.9.0.80
python-telegram-bot==20.8

# https://download.pytorch.org/whl/cpu/torch-1.10.2%2Bcpu-cp39-cp39-linux_x86_64.whl
# https://download.pytorch.org/whl/cpu/torch-1.10.2-cp38-cp38-manylinux2014_aarch64.whl